make update work from menu
